Conflicts:
	.gitmodules

+++ platform/gl/gl-font.c
@@ -1,3 +1,25 @@
+// Copyright (C) 2004-2021 Artifex Software, Inc.
+//
+// This file is part of MuPDF.
+//
+// MuPDF is free software: you can redistribute it and/or modify it under the
+// terms of the GNU Affero General Public License as published by the Free
+// Software Foundation, either version 3 of the License, or (at your option)
+// any later version.
+//
+// MuPDF is distributed in the hope that it will be useful, but WITHOUT ANY
+// WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
+// details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with MuPDF. If not, see <https://www.gnu.org/licenses/agpl-3.0.en.html>
+//
+// Alternative licensing terms are available from the licensor.
+// For commercial licensing, see <https://www.artifex.com/> or contact
+// Artifex Software, Inc., 1305 Grant Avenue - Suite 200, Novato,
+// CA 94945, U.S.A., +1(415)492-9861, for further information.
+
 /*
  * A very simple font cache and rasterizer that uses FreeType
  * to draw fonts from a single OpenGL texture. The code uses
@@ -61,8 +83,8 @@ static fz_font *g_font = NULL;
 static void clear_font_cache(void)
 {
 #if PADDING > 0
-	unsigned char *zero = malloc(g_cache_w * g_cache_h);
-	memset(zero, 0, g_cache_w * g_cache_h);
+	unsigned char *zero = malloc((size_t)g_cache_w * g_cache_h);
+	memset(zero, 0, (size_t)g_cache_w * g_cache_h);
 	glBindTexture(GL_TEXTURE_2D, g_cache_tex);
 	glTexSubImage2D(GL_TEXTURE_2D, 0, 0, 0, g_cache_w, g_cache_h, GL_ALPHA, GL_UNSIGNED_BYTE, zero);
 	free(zero);
